Move X-Cube decoder to decoders/xcube/ folder (current version: project in all planes)
