+++ panqec/decoders/__init__.py
@@ -4,7 +4,7 @@ from .bposd.bposd_decoder import BeliefPropagationOSDDecoder  # noqa
 from .bposd.mbp_decoder import MemoryBeliefPropagationDecoder  # noqa
 
 from .matching._matching_decoder import MatchingDecoder  # noqa
-from .matching._xcube_matching_decoder import XCubeMatchingDecoder  # noqa
+from .xcube._xcube_matching_decoder import XCubeMatchingDecoder  # noqa
 from .sweepmatch._sweep_decoder_3d import SweepDecoder3D  # noqa
 from .sweepmatch._sweep_match_decoder import SweepMatchDecoder  # noqa
 from .sweepmatch._rotated_sweep_decoder import RotatedSweepDecoder3D  # noqa
